rename circuit
